🎨 Refactor test datasets
