+++ lamindb/core/datasets/_core.py
@@ -342,33 +342,6 @@ def anndata_human_immune_cells(
     return adata
 
 
-def anndata_with_obs() -> ad.AnnData:
-    """Create a mini anndata with cell_type, disease and tissue."""
-    import anndata as ad
-    import bionty.base as bionty_base
-
-    celltypes = ["T cell", "hematopoietic stem cell", "hepatocyte", "my new cell type"]
-    celltype_ids = ["CL:0000084", "CL:0000037", "CL:0000182", ""]
-    diseases = [
-        "chronic kidney disease",
-        "liver lymphoma",
-        "cardiac ventricle disorder",
-        "Alzheimer disease",
-    ]
-    tissues = ["kidney", "liver", "heart", "brain"]
-    df = pd.DataFrame()
-    df["cell_type"] = celltypes * 10
-    df["cell_type_id"] = celltype_ids * 10
-    df["tissue"] = tissues * 10
-    df["disease"] = diseases * 10
-    df.index = "obs" + df.index.astype(str)
-
-    adata = ad.AnnData(X=np.zeros(shape=(40, 100), dtype=np.float32), obs=df)
-    adata.var.index = bionty_base.Gene().df().head(100)["ensembl_gene_id"].values
-
-    return adata
-
-
 def anndata_suo22_Visium10X():  # pragma: no cover
     """AnnData from Suo22 generated by 10x Visium."""
     import anndata as ad
